backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
email-validator
passlib[bcrypt]
bcrypt
python-jose
python-multipart
yfinance>=1.2.0
openai
